# perf: RPC Serialization