… go/vt/proto/automation/automation.pb.go
… go/vt/proto/automationservice/automationservice.pb.go
… go/vt/proto/binlogdata/binlogdata.pb.go
… go/vt/proto/binlogservice/binlogservice.pb.go
… go/vt/proto/logutil/logutil.pb.go
… go/vt/proto/mysqlctl/mysqlctl.pb.go
… go/vt/proto/query/query.pb.go
… go/vt/proto/queryservice/queryservice.pb.go
… go/vt/proto/replicationdata/replicationdata.pb.go
… go/vt/proto/tableacl/tableacl.pb.go
… go/vt/proto/tabletmanagerdata/tabletmanagerdata.pb.go
… go/vt/proto/tabletmanagerservice/tabletmanagerservice.pb.go
… go/vt/proto/throttlerdata/throttlerdata.pb.go
… go/vt/proto/throttlerservice/throttlerservice.pb.go
… go/vt/proto/topodata/topodata.pb.go
… go/vt/proto/vschema/vschema.pb.go
… go/vt/proto/vtadmin/vtadmin.pb.go
… go/vt/proto/vtctldata/vtctldata.pb.go
… go/vt/proto/vtctlservice/vtctlservice.pb.go
… go/vt/proto/vtgate/vtgate.pb.go
… go/vt/proto/vtgateservice/vtgateservice.pb.go
… go/vt/proto/vtrpc/vtrpc.pb.go
… go/vt/proto/vttest/vttest.pb.go
… go/vt/proto/vttime/vttime.pb.go
… go/vt/proto/vtworkerdata/vtworkerdata.pb.go
… go/vt/proto/vtworkerservice/vtworkerservice.pb.go
… go/vt/proto/workflow/workflow.pb.go