@derekperkins 2021-02-19
I'm not sure that there's any performance gain worth switching to gogo. All of the choices by major projects to use it was prior to the new api, and they haven't changed yet simply because of the work required to change.

- The project has been almost completely abandoned for nearly a year, with nobody willing to commit to taking over. https://github.com/gogo/protobuf/issues/691
- etcd using gogo has caused an enormous amount of pain for anyone using it as a dependency https://github.com/etcd-io/etcd/issues/12124
- Kubernetes is looking to migrate away https://github.com/kubernetes/kubernetes/issues/96564 https://groups.google.com/g/kubernetes-sig-api-machinery/c/tcwFubV9Boo/m/MAdQfLY_AgAJ
- Anyone using our Go database driver (we're one) is actively prevented from upgrading to newer versions of grpc-go because of that etcd dependency, and picking this incompatible fork is likely to cause more issues in the future

I don't have a personal opinion about the new apis, though you're not the first person I've talked to with a similar opinion. I just feel like this is backing ourselves into a corner that we are going to seriously regret later.