@vmg 2021-02-22
> I'm not sure that there's any performance gain worth switching to gogo.

The unfortunate truth is that we're not going to find another 3% global throughput increase in Vitess without a massive engineering effort. That engineering effort could be scoped to other parts of the codebase, or it could be scoped to ensuring Gogo (or an equivalent) is properly maintained in the long term. The metrics show that our messages in production are actually quite large _and_ complex, and they need to be optimized because they're showing up in the graphs.

How do we accomplish that in the most sustainable way? Well, my recommendation for this PR is going to be to use the backwards compatible Gogo generator, which doesn't introduce any actual dependencies on the unmaintained Gogo packages (even though, unfortunately, it's not the fastest). I think it's a good middle ground that gets us unrolled, faster code, while still ensuring we're depending on Google's APIs.

Once we have removed this low-hanging fruit, I intend to research how hard it'd be to wire up the generation to the new ProtoBuf API v2. It appears that the `gogo` project provides a lot of functionality which we don't intend to use and which would be very hard to port to the new API v2, but if we could depend on our own generator, which would be implemented as a plug-in to Google's `protoc-gen-go`, as opposed to `gogo`'s approach of replacing the generator altogether, that seems like a good middle point for performance and maintainability -- again, always judging by the graphs showed in this PR.