@sougou 2021-02-22
@vmg This is awesome research! :+1: 

Someone also told me that the http2 transport layer under grpc could be a CPU hog. It's another thing to look at, at least to confirm/deny the theory.