@derekperkins 2021-02-22
I don't want to come across too negative about the potential performance gains, I think this is awesome. I just want to make sure we're going in with the understanding that gogoproto has been without maintainers for a year, and while this PR might be scoped cleanly, it's all too easy to have future code depend on specific gogo behavior that would be near impossible to back out.

> if we could depend on our own generator, which would be implemented as a plug-in to Google's protoc-gen-go, as opposed to gogo's approach of replacing the generator altogether, that seems like a good middle point for performance and maintainability

This sounds like a good path forward. Given that we only need a subset of functionality and that the transport consumes so many resources, it seems reasonable to take control of that piece, and as you mentioned at the beginning, we could even explore non-protobuf protocols if data showed it was worth the effort.