@vmg 2021-02-23
:+1: I definitely agree that depending on the `gogo/protobuf` package directly is too risky. I've refactored the PR so we only depend on the codegen and we remain fully backwards compatible. It's ready for review now!