@vmg 2021-02-24
@shlomi-noach it's already linted! I've forced the usage of the backwards-compatible generator, so any `gogo`-exclusive features that are introduced will just crash the build. :ok_hand: