@ajm188 2021-02-24
@vmg I'm getting the following running `make proto` after this is merged.

```
vtctldata.proto:167:3: "google.protobuf.Duration" is not defined.
vtctldata.proto:334:3: "google.protobuf.Duration" is not defined.
```

Do you happen to know a workaround for this?