@vmg 2021-02-25
Yes! You got it! We need to use our own `vttime.Duration` to ensure we can generate the optimized code for it. It's fully compatible with the default Duration in ProtoBufs